maps/dictionary.go
package maps

import "errors"

type Dictionary map[string]string

var (
	ErrorWordNotFound      = errors.New("expected to find word")
	ErrorWordExists        = errors.New("word already exists")
	ErrorWordDoesNotExists = errors.New("word does not exists")
)

func (d Dictionary) Search(word string) (string, error) {
	definition, found := d[word]
	if !found {
		return "", ErrorWordNotFound
	}

	return definition, nil
}

func (d Dictionary) Add(word, definition string) error {
	_, err := d.Search(word)

	switch err {
	case ErrorWordNotFound:
		d[word] = definition
	case nil:
		return ErrorWordExists
	default:
		return err
	}

	return nil
}

func (d Dictionary) Update(word, definition string) error {
	_, err := d.Search(word)

	switch err {
	case ErrorWordNotFound:
		return ErrorWordDoesNotExists
	case nil:
		d[word] = definition
	default:
		return err
	}

	return nil
}

func (d Dictionary) Delete(word string) {
	delete(d, word)
}
